refactor(component)!: move utils/createNamespace to components/namespace
BREAKING CHANGE: the namespace is now created through import { createNamespace } from "@socialgouv/kosko-charts/components/namespace"

```diff
- import { createNamespace } from "@socialgouv/kosko-charts/utils/createNamespace";
+ import { createNamespace } from "@socialgouv/kosko-charts/components/namespace";
```
