+++ templates/simple/components/_namespace.ts
@@ -1,4 +1,4 @@
-import { createNamespace } from "@socialgouv/kosko-charts/utils/createNamespace";
+import { createNamespace } from "@socialgouv/kosko-charts/components/namespace";
 
 const manifests = createNamespace();
